requirements.txt
numpy>=1.21.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.7.0
Pillow>=9.0.0
tqdm>=4.65.0
flask>=2.0.0
flask-socketio>=5.3.0
flask-cors>=3.0.10
ultralytics>=8.0.0
facenet-pytorch>=2.5.0
transformers>=4.30.0
open-clip-torch>=2.20.0
faiss-cpu>=1.7.0
h5py>=3.8.0
gfpgan>=1.3.8
python-socketio>=5.7.0
eventlet>=0.33.0
